Allow use of local appimagetool in build

+++ scripts/helpers/functions.sh
@@ -306,7 +306,7 @@ generate_type2_appimage()
   #   URL=$(wget -q "https://s3.amazonaws.com/archive.travis-ci.org/jobs/$((ID+2))/log.txt" -O - | grep "https://transfer.sh/.*/appimagetool" | tail -n 1 | sed -e 's|\r||g')
   # fi
   URL="https://github.com/AppImage/AppImageKit/releases/download/continuous/appimagetool-${SYSTEM_ARCH}.AppImage"
-  wget -c "$URL" -O appimagetool
+  [ -e appimagetool ] || wget -c "$URL" -O appimagetool
   chmod a+x ./appimagetool
   appimagetool=$(readlink -f appimagetool)
